requirements.txt
langgraph>=1.0.5
langchain>=1.2.0
langchain-core>=1.2.1
langchain-groq>=1.1.0
chromadb>=1.5.0
langchain-chroma>=1.1.0
sentence-transformers>=5.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0
